docs(manifest): add `homepage_url`i
Adds the project's GitHub repository URL as the `homepage_url` to the
`manifest.json`, so that it is displayed in the Firefox / Chrome
extension store listing. This makes it easier for users to find the
source code and contribute fixes or features.

`homepage_url` is universally understood by Firefox, as well as the
Chrome Extensions Manifest V3 & V2, which is what's currently used by
this project.

https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/homepage_url
https://developer.chrome.com/docs/extensions/mv3/manifest/
https://developer.chrome.com/docs/extensions/mv2/manifest/